+++ cli/src/utils/log.rs
@@ -2,7 +2,7 @@ use std::env;
 
 /// Initialize the logger.
 pub fn initialize_logger() {
-	if let Err(_) = env::var("RUST_LOG") {
+	if env::var("RUST_LOG").is_err() {
 		env::set_var("RUST_LOG", "info");
 	}
 	pretty_env_logger::init();