+++ CHANGELOG.md
@@ -15,8 +15,9 @@ new versions as soon as we feel it is necessary.
 
 ## Unreleased
 
-- Parallel package vendoring (#13)
-- Split common module into separate files (#14)
+- Parallel package vendoring ([#13](https://github.com/nix-community/mineflake/pull/13))
+- Split common module into separate files ([#14](https://github.com/nix-community/mineflake/pull/14))
+- Fix CHANGELOG.md links to PRs ([#15](https://github.com/nix-community/mineflake/pull/15))
 
 ## 0.2.1 - 15 Dec 2022
 
@@ -27,7 +28,7 @@ new versions as soon as we feel it is necessary.
 - **Repository moved to nix-community GitHub organization**
 - Updated contribution guides ([#57](https://git.frsqr.xyz/firesquare/mineflake/pulls/57))
 - Relicense from GPL3 to MIT ([#57](https://git.frsqr.xyz/firesquare/mineflake/pulls/57))
-- Rewrite main logic from Nix to Rust (#9) **BREAKING CHANGE**
+- Rewrite main logic from Nix to Rust ([#9](https://github.com/nix-community/mineflake/pull/9)) **BREAKING CHANGE**
 
 ## 0.1 - 2 Sep 2022